+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Add `Signal#switchMap` method
 * Fix an issue where `concatMap` wouldn't wait for signals to complete
 * Add missing combinators to exports
+* Add `concat` function
+* Add `Signal#concat` method
 
 ## 1.4.0 (2019-01-08)